Improve generic DB creation error reporting
This moves the reporting of errors that occurred in a reader thread to just after the thread completes.
Doing this avoids dropping an error when trying to add together arrays of different shapes if a reader thread errored because the feature/label shape was not constant.

Assertions are replaced with more meaningful exceptions (helps display the real error in DIGITS UI i.e. instead of showing `AssertionError` we are now showing `ValueError: shape mismatch ...`).